Refactor auth package structure for better organization
Reorganize the authentication and authorization code into a cleaner,
more maintainable package structure with better separation of concerns.

Key changes:
- Extract OIDC discovery into dedicated pkg/auth/oidc package
- Split token validation and introspection into pkg/auth/token with
  provider pattern for different introspection endpoints
- Move HTTP middleware to pkg/auth/middleware package
- Extract PKCE generation into separate file in oauth package
- Consolidate RFC 9728 resource metadata handling
- Add backward compatibility layer in pkg/auth/compat.go

The refactoring reduces code duplication, improves testability, and
maintains full backward compatibility through deprecated type aliases
and wrapper functions.

Net reduction: 3,394 lines of code (-65% in modified files)
