+++ pkg/auth/discovery/discovery.go
@@ -10,9 +10,7 @@ package discovery
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
 	"net/url"
 	"path"
@@ -21,7 +19,6 @@ import (
 
 	"golang.org/x/oauth2"
 
-	"github.com/stacklok/toolhive/pkg/auth"
 	"github.com/stacklok/toolhive/pkg/auth/oauth"
 	"github.com/stacklok/toolhive/pkg/logger"
 	"github.com/stacklok/toolhive/pkg/networking"
@@ -156,83 +153,6 @@ func detectAuthWithRequest(
 	return nil, nil
 }
 
-// ParseWWWAuthenticate parses the WWW-Authenticate header to extract authentication information
-// Supports multiple authentication schemes and complex header formats
-func ParseWWWAuthenticate(header string) (*AuthInfo, error) {
-	// Trim whitespace and handle empty headers
-	header = strings.TrimSpace(header)
-	if header == "" {
-		return nil, fmt.Errorf("empty WWW-Authenticate header")
-	}
-
-	// Check for OAuth/Bearer authentication
-	// Note: We don't split by comma because Bearer parameters can contain commas in quoted values
-	if strings.HasPrefix(header, "Bearer") {
-		authInfo := &AuthInfo{Type: "OAuth"}
-
-		// Extract parameters after "Bearer"
-		params := strings.TrimSpace(strings.TrimPrefix(header, "Bearer"))
-		if params != "" {
-			// Parse parameters (realm, scope, resource_metadata, etc.)
-			realm := ExtractParameter(params, "realm")
-			if realm != "" {
-				authInfo.Realm = realm
-			}
-
-			// RFC 9728: Check for resource_metadata parameter
-			resourceMetadata := ExtractParameter(params, "resource_metadata")
-			if resourceMetadata != "" {
-				authInfo.ResourceMetadata = resourceMetadata
-			}
-
-			// Extract error information if present
-			errorParam := ExtractParameter(params, "error")
-			if errorParam != "" {
-				authInfo.Error = errorParam
-			}
-
-			errorDesc := ExtractParameter(params, "error_description")
-			if errorDesc != "" {
-				authInfo.ErrorDescription = errorDesc
-			}
-		}
-
-		return authInfo, nil
-	}
-
-	// Check for OAuth-specific schemes
-	if strings.HasPrefix(header, "OAuth") {
-		authInfo := &AuthInfo{Type: "OAuth"}
-
-		// Extract parameters after "OAuth"
-		params := strings.TrimSpace(strings.TrimPrefix(header, "OAuth"))
-		if params != "" {
-			// Parse parameters (realm, scope, etc.)
-			realm := ExtractParameter(params, "realm")
-			if realm != "" {
-				authInfo.Realm = realm
-			}
-
-			// RFC 9728: Check for resource_metadata parameter
-			resourceMetadata := ExtractParameter(params, "resource_metadata")
-			if resourceMetadata != "" {
-				authInfo.ResourceMetadata = resourceMetadata
-			}
-		}
-
-		return authInfo, nil
-	}
-
-	// Currently only OAuth-based authentication is supported
-	// Basic and Digest authentication are not implemented
-	if strings.HasPrefix(header, "Basic") || strings.HasPrefix(header, "Digest") {
-		logger.Debugf("Unsupported authentication scheme: %s", header)
-		return nil, fmt.Errorf("unsupported authentication scheme: %s", strings.Split(header, " ")[0])
-	}
-
-	return nil, fmt.Errorf("no supported authentication type found in header: %s", header)
-}
-
 // DeriveIssuerFromURL attempts to derive the OAuth issuer from the remote URL using general patterns
 func DeriveIssuerFromURL(remoteURL string) string {
 	// Parse the URL to extract the domain
@@ -266,57 +186,6 @@ func DeriveIssuerFromURL(remoteURL string) string {
 
 	logger.Debugf("Derived issuer from URL - remoteURL: %s, issuer: %s", remoteURL, issuer)
 	return issuer
-}
-
-// ExtractParameter extracts a parameter value from an authentication header
-// Handles both quoted and unquoted values according to RFC 2617 and RFC 6750
-func ExtractParameter(params, paramName string) string {
-	// Parameters can be separated by comma or space
-	// Handle both paramName=value and paramName="value" formats
-
-	// First try to find the parameter with equals sign
-	searchStr := paramName + "="
-	idx := strings.Index(params, searchStr)
-	if idx == -1 {
-		return ""
-	}
-
-	// Extract the value after the equals sign
-	valueStart := idx + len(searchStr)
-	if valueStart >= len(params) {
-		return ""
-	}
-
-	remainder := params[valueStart:]
-
-	// Check if the value is quoted
-	if strings.HasPrefix(remainder, `"`) {
-		// Find the closing quote
-		endIdx := 1
-		for endIdx < len(remainder) {
-			if remainder[endIdx] == '"' && (endIdx == 1 || remainder[endIdx-1] != '\\') {
-				// Found unescaped closing quote
-				value := remainder[1:endIdx]
-				// Unescape any escaped quotes
-				value = strings.ReplaceAll(value, `\"`, `"`)
-				return value
-			}
-			endIdx++
-		}
-		// No closing quote found, return empty
-		return ""
-	}
-
-	// Unquoted value - find the end (comma, space, or end of string)
-	endIdx := 0
-	for endIdx < len(remainder) {
-		if remainder[endIdx] == ',' || remainder[endIdx] == ' ' {
-			break
-		}
-		endIdx++
-	}
-
-	return strings.TrimSpace(remainder[:endIdx])
 }
 
 // DeriveIssuerFromRealm attempts to derive the OAuth issuer from the realm parameter
@@ -544,71 +413,6 @@ func registerDynamicClient(
 	}
 
 	return registrationResponse, nil
-}
-
-// FetchResourceMetadata as specified in RFC 9728
-func FetchResourceMetadata(ctx context.Context, metadataURL string) (*auth.RFC9728AuthInfo, error) {
-	if metadataURL == "" {
-		return nil, fmt.Errorf("metadata URL is empty")
-	}
-
-	// Validate URL
-	parsedURL, err := url.Parse(metadataURL)
-	if err != nil {
-		return nil, fmt.Errorf("invalid metadata URL: %w", err)
-	}
-
-	// RFC 9728: Must use HTTPS (except for localhost in development)
-	if parsedURL.Scheme != "https" && parsedURL.Hostname() != "localhost" && parsedURL.Hostname() != "127.0.0.1" {
-		return nil, fmt.Errorf("metadata URL must use HTTPS: %s", metadataURL)
-	}
-
-	// Create HTTP client with timeout
-	client := &http.Client{
-		Timeout: DefaultHTTPTimeout,
-		Transport: &http.Transport{
-			TLSHandshakeTimeout:   5 * time.Second,
-			ResponseHeaderTimeout: 5 * time.Second,
-		},
-	}
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, metadataURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	req.Header.Set("Accept", "application/json")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed to fetch metadata: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("metadata request failed with status %d", resp.StatusCode)
-	}
-
-	// Check content type
-	contentType := strings.ToLower(resp.Header.Get("Content-Type"))
-	if !strings.Contains(contentType, "application/json") {
-		return nil, fmt.Errorf("unexpected content type: %s", contentType)
-	}
-
-	// Parse the metadata
-	const maxResponseSize = 1024 * 1024 // 1MB limit
-	var metadata auth.RFC9728AuthInfo
-	if err := json.NewDecoder(io.LimitReader(resp.Body, maxResponseSize)).Decode(&metadata); err != nil {
-		return nil, fmt.Errorf("failed to parse metadata: %w", err)
-	}
-
-	// RFC 9728 Section 3.3: Validate that the resource value matches
-	// For now we just check it's not empty
-	if metadata.Resource == "" {
-		return nil, fmt.Errorf("metadata missing required 'resource' field")
-	}
-
-	return &metadata, nil
 }
 
 // ValidateAndDiscoverAuthServer attempts to validate if a URL is an authorization server
