+++ pkg/auth/oauth/dynamic_registration.go
@@ -155,7 +155,7 @@ func createHTTPRequest(
 }
 
 // getHTTPClient returns the HTTP client to use for the request
-func getHTTPClient(client httpClient) httpClient {
+func getHTTPClient(client *http.Client) *http.Client {
 	if client != nil {
 		return client
 	}
@@ -210,7 +210,7 @@ func registerClientDynamicallyWithClient(
 	ctx context.Context,
 	registrationEndpoint string,
 	request *DynamicClientRegistrationRequest,
-	client httpClient,
+	client *http.Client,
 ) (*DynamicClientRegistrationResponse, error) {
 	// Validate registration endpoint URL
 	if _, err := validateRegistrationEndpoint(registrationEndpoint); err != nil {
